backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-storages==1.14.2
celery==5.3.4
redis==5.0.1
requests==2.31.0
geopy==2.4.1